refactor: transaction-signing
