+++ src/pages/transaction-signing/components/errors/transaction-error.spec.tsx
