+++ src/pages/transaction-signing/components/fee-row/fee.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { stacksValue } from '@common/stacks-utils';
 import { useTransactionFee } from '@pages/transaction-signing/hooks/use-transaction-fee';
 
-export const FeeComponent = () => {
+export const Fee = () => {
   const { isSponsored, amount } = useTransactionFee();
   if (typeof amount === 'undefined') return null;
   return (
