+++ src/pages/transaction-signing/components/speed-row/fee-estimate-dropdown.tsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export const FeeEstimateDropdown = () => {
+  return <>Fee estimate dropdown goes here...</>;
+};
